pyproject.toml
[project]
name = "rag-blueprint-vespa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.0.1",
    "lightgbm>=4.6.0",
    "pandas>=2.3.0",
    "pyvespa>=0.57.0",
    "scikit-learn>=1.7.0",
]
